chore: cleanup CacheHandler factoring
